public credentials test
