+++ pallets/public-credentials/src/tests/mod.rs
@@ -16,29 +16,7 @@
 
 // If you feel like getting in touch with us, you can do so at info@botlabs.org
 
-use frame_support::{assert_noop, assert_ok, crypto::ecdsa::ECDSAExt, traits::fungible::InspectHold};
-use kilt_support::{mock::mock_origin, Deposit};
-use parity_scale_codec::Encode;
-use sha3::{Digest, Keccak256};
-use sp_runtime::{
-	app_crypto::{ecdsa, sr25519, Pair},
-	traits::{IdentifyAccount, Zero},
-	MultiSignature, MultiSigner, TokenError,
-};
-
-use crate::{
-	account::{AccountId20, EthereumSignature},
-	associate_account_request::{get_challenge, AssociateAccountRequest},
-	linkable_account::LinkableAccountId,
-	mock::*,
-	signature::get_wrapped_payload,
-	ConnectedAccounts, ConnectedDids, ConnectionRecord, Error, HoldReason,
-};
-
-
-
-
-
- 
-
-
+mod create;
+mod delete;
+mod deposit;
+mod revoke;
